requirements.txt
Django==2.2
djangorestframework==3.9.3
pytz==2019.1
sqlparse==0.3.0
